// src/managed/staging/delivery-env.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types src/managed/staging/delivery-env.d.ts -c managed/staging/delivery.json --env staging --strict-vars false --env-interface StagingDeliveryEnv --include-runtime false` (hash: 1a0cc130097240ce7320aade4d1fd6df)
interface __BaseEnv_StagingDeliveryEnv {
  ENVIRONMENT: string;
  STAGING_ENABLED: string;
  STAGING_DELIVERY_ENABLED: string;
  STAGING_RECEIPTS: DurableObjectNamespace<import('./delivery').StagingReceipt>;
  STAGING_AUTHORITY: Service /* entrypoint DeliveryAuthority from bugdrop-managed-authority-staging */;
  STAGING_GITHUB: Fetcher /* bugdrop-managed-github-staging */;
}
declare namespace Cloudflare {
  interface GlobalProps {
    mainModule: typeof import('./delivery');
    durableNamespaces: 'StagingReceipt';
  }
  interface Env extends __BaseEnv_StagingDeliveryEnv {}
}
interface StagingDeliveryEnv extends __BaseEnv_StagingDeliveryEnv {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
  [Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
  interface ProcessEnv extends StringifyValues<
    Pick<Cloudflare.Env, 'ENVIRONMENT' | 'STAGING_ENABLED' | 'STAGING_DELIVERY_ENABLED'>
  > {}
}

export { StagingDeliveryEnv };